+++ src/multicolor/passband/mod.rs
@@ -7,5 +7,8 @@ pub use dump_passband::DumpPassband;
 mod passband_trait;
 pub use passband_trait::PassbandTrait;
 
+mod labeled_passband;
+pub use labeled_passband::LabeledPassband;
+
 mod string_passband;
 pub use string_passband::StringPassband;